requirements.txt
torch>=1.13.0
torchvision>=0.14.0
onnx>=1.12.0
onnxruntime>=1.14.0
timm>=0.9.2
numpy
opencv-python
Pillow
tqdm
scipy
matplotlib

pycuda         
tensorrt 
